requirements.txt
Flask
Flask-Cors
tensorflow
opencv-python
numpy
Pillow
gunicorn
